src/requirements.txt
fastapi==0.110.2
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
pydantic-settings==2.2.1
aiofiles==23.2.1
langchain==0.1.20
langchain-community==0.0.38
PyMuPDF==1.23.26
motor==3.7.1
openai==2.26.0
google-genai==1.68.0
cohere==5.20.7
qdrant-client==1.17.0
ollama==0.6.1
SQLAlchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0